Chef cookbook to provision a win. build machine
WHich can also be a Vagrant VM. This one could actually be merged witht
he OSX one, it would imply a couple of ifs and the `provision.sh` script
wouldn't be needed for windows since we can use vagrant-berkshelf.

One of the other improvements to bring would be to remove a couple of
our omnibus software DSLs in `omnibus-software` (libgcc, make...) for
instance and have them being installed by the cookbook. They don't need
to be shipped with the agent so there's no reason why they would be
built by Omnibus.